cab-booking
1	Introduction
Auto Taxi booking system is a software application that allows users to book and schedule rides with a taxi service. The system provides a platform for customers to request a ride, track their vehicle in real time, and make payments securely. With an auto taxi booking system, passengers can enjoy a convenient, hassle-free, and efficient transportation experience.

2	Motivation

	


3	Objectives

Increase convenience for customers.
Bring together drivers and riders.
Provide a safe and reliable transportation option.
Improve the transportation experience for everyone.


4	Functionalities
.

5	Developing Environment

This application is developed using Django, a Python-based web framework, to efficiently structure and manage the project. Key Django components include models for representing data (products, stores, and users), views for handling user interactions, and templates for creating the user interface.  For user authentication and authorization, Django's built-in authentication system is employed. To integrate mapping functionalities and display store locations, the Google Maps API is utilized

6	Conclusion

In conclusion, the auto-taxi booking system project is a valuable addition to the transportation industry. The system provides a convenient way for customers to book a taxi or an auto-rickshaw from anywhere at any time. The project has several benefits such as reducing the waiting time, increasing the efficiency of taxi or auto-rickshaw drivers, and providing a safe and secure mode of transportation for passengers. The project’s success depends on several factors such as the availability and reliability of the internet, and the responsiveness of the drivers. The project also requires a robust backend system to handle booking requests, payment processing, and driver tracking. Overall, the auto-taxi booking system project has the potential to revolutionize the transportation industry by providing an efficient, reliable, and affordable transportation option for millions of people worldwide